+++ core/compare.py
@@ -1,14 +1,18 @@
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
 import shutil
 import glob
-from core.storage import misc
-from core import ingest
+import itertools
+import matplotlib.pyplot as plt
+from functools import reduce
+
+import numpy as np
+import pandas as pd
 from rdkit import Chem
 from rdkit.Chem import PandasTools
-import itertools
-from functools import reduce
+
+from core.ingest import resolveID
+from core.storage import cd
+
+
 
 """
 Goal is to compare the output of many models in a systematic, automatic manner.
@@ -21,7 +25,7 @@ def datasize(dataset):
     Flaw is the assumption that each row is a valid data set
     entry.  It is safe for our basic cases.
     """
-    with misc.cd('../dataFiles/'): # move to dataset directory
+    with cd('../dataFiles/'): # move to dataset directory
         with open(dataset) as f:
             size = sum(1 for line in f) - 1  # remove one for header
             print('The {} dataset has {} entries in it.'.format(dataset, size))
@@ -329,7 +333,7 @@ def moloverlap(datasets, n, image=False):
     image=False,  Whether to create a grid image of the overlapping molecules.
 
     """
-    with misc.cd('../dataFiles/'):  # move to dataset directory
+    with cd('../dataFiles/'):  # move to dataset directory
 
         # use dictionary comprehension and iterate over argument dictionary
         # create resolve chemID and create dataframe with ingest.resolveID
@@ -345,7 +349,7 @@ def moloverlap(datasets, n, image=False):
             PandasTools.AddMoleculeColumnToFrame(df, col, 'Molecule', includeFingerprints=True)
             if df['Molecule'].isnull().values.any():  # if any ID failed to be converted to molobjects
                 print(dataset, 'failed to create objects. Attempting resolve.')
-                df = ingest.resolveID(dataset, col) # try to find smiles for the ID column
+                df = resolveID(dataset, col) # try to find smiles for the ID column
                 # retry converting to molobject after resolution
                 PandasTools.AddMoleculeColumnToFrame(df, 'smiles', 'Molecule', includeFingerprints=True)
