+++ core/neo4j/output_to_neo4j.py
@@ -2,14 +2,15 @@
 Objective: Contains the function "output_to_neo4j" which is the function that imports data from output folders into
             Neo4j based in our ontology
 """
-import pandas as pd
 import os
-from zipfile import ZipFile
-from core.storage import misc
 import json
-from core.neo4j import prep_from_outputs
-import pathlib
+from zipfile import ZipFile
 from main import ROOT_DIR
+
+import pandas as pd
+
+from core.storage import cd
+from core.neo4j.prep_from_outputs import Prep
 
 
 def file_count():
@@ -22,7 +23,7 @@ def file_count():
     """
     # print(hello)
     os.chdir(ROOT_DIR)  # Start in root directory
-    with misc.cd('output'):  # Access output
+    with cd('output'):  # Access output
         print('Now in:', os.getcwd())
         for roots, dirs, files in os.walk(os.getcwd()):
             files_count = 0
@@ -46,7 +47,7 @@ def get_file(file_string):
     :return: 3 dataframe from _attributes.json, _data.csv and _predictions.csv
     """
     os.chdir(ROOT_DIR)  # Start in root directory
-    with misc.cd('output'):  # Access output
+    with cd('output'):  # Access output
         for roots, dirs, files in os.walk(os.getcwd()):
             for f in files:
                 if f.endswith('.zip'):
@@ -83,7 +84,7 @@ def output_to_neo4j(port, username, password):
         # pandas dataframe
         df_from_data = next(data_csv_generator)
         df_from_predictions = next(predictions_csv_generator)
-        prep = prep_from_outputs.Prep(df_from_attributes, df_from_predictions, df_from_data)
+        prep = Prep(df_from_attributes, df_from_predictions, df_from_data)
         prep.to_neo4j(port=port, username=username, password=password)
